+++ src/frontend/src/pages/FinancePage/Finance.tsx
@@ -10,7 +10,7 @@ import EditReimbursementRequestPage from './EditReimbursementRequest/EditReimbur
 import ReimbursementRequestDetails from './ReimbursementRequestDetailPage/ReimbursementRequestDetails';
 import FinanceDashboard from './FinanceDashboard';
 import ReimbursmentRequests from './ReimbursmentRequests';
-import VendorsAndSponsorsPage from './VendorsAndSponsorsPage';
+import CompaniesPage from './CompaniesPage';
 
 const Finance: React.FC = () => {
   return (
@@ -20,7 +20,7 @@ const Finance: React.FC = () => {
       <Route path={routes.REIMBURSEMENT_REQUEST_BY_ID} component={ReimbursementRequestDetails} />
       <Route path={routes.FINANCE_DASHBOARD} component={FinanceDashboard} />
       <Route path={routes.REIMBURSEMENT_REQUESTS} component={ReimbursmentRequests} />
-      <Route path={routes.COMPANIES} component={VendorsAndSponsorsPage} />
+      <Route path={routes.COMPANIES} component={CompaniesPage} />
       <Route path={routes.FINANCE} component={FinancePage} />
     </Switch>
   );
